src/crawler/requirements.txt
attrs==21.2.0
Automat==20.2.0
autopep8==1.5.7
backports.zoneinfo==0.2.1
base==1.0.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
constantly==15.1.0
cryptography==3.4.8
cssselect==1.1.0
docopt==0.6.2
greenlet==1.1.2
grpcio==1.41.0
grpcio-tools==1.41.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
pipreqs==0.4.10
priority==1.3.0
Protego==0.1.16
protobuf==3.18.0
psycopg==3.0b1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
queuelib==1.6.2
requests==2.26.0
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
Twisted==21.7.0
typing-extensions==3.10.0.2
urllib3==1.26.7
w3lib==1.22.0
yarg==0.1.9
zope.interface==5.4.0
